add tasks to worker in active
